requirements.txt
# Core Dependencies
# Requires Python 3.10 or higher

# LLM & RAG Orchestration
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Vector Database
chromadb>=0.4.22

# Ontology & Reasoning (CRITICAL: Must use HermiT reasoner)
owlready2>=0.45

# Document Processing
pypdf>=3.17.0
tiktoken>=0.5.0

# Data Processing
numpy>=1.26.0
pandas>=2.0.0

# OpenAI API
openai==1.7.2

# Optional: Local LLM Support (Ollama)
# ollama==0.1.6

# HTTP Requests (for downloading FIBO files)
requests==2.31.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing & Development
pytest==7.4.3
black==23.12.1
